Introducing broadcasting nodes
In order to reduce the amount of connections each node needs to maintain in the case of broadcasting alerts, make it possible to specify a list of endpoints that will act as broadcasting nodes.
Each non-broadcast node will send a broadcasting message to the broadcasting node which contains the message to broadcast and the recipients. The broadcasting node is selected using consistent hashing
so as to keep the amount of connections of the non-broadcasting nodes to a minimum and to evenly distribute load amongst the broadcasting nodes.