Anti-entropy mechanism
It can happen that a node misses a part of the consensus messages whilst still being able to send out its own vote (unidirectional network partition, message overload, ...). In this case, the rest of the group will see this node as being part of the group and the monitoring mechanism will still be working as expected, but the stale node will run an old configuration.

In order to enforce consistency in this case, the following new anti-entropy mechanism is used:

- each node maintains a set of configurations it has been part of
- probe messages now contain the configuration ID of the observer
- when a node receives a probe message with a configuration ID it does not know, it will start a background task to check again after a configured timeout (1 minute by default)
- if the configuration ID is still unknown after the timeout has reached, the node leaves (using the LEAVE protocol)